rhs sidebar works for markdoc
