+++ docs/next/pages/[...page].tsx
@@ -5,7 +5,7 @@ import zlib from 'zlib';
 
 import FeedbackModal from 'components/FeedbackModal';
 import {Shimmer} from 'components/Shimmer';
-import {getItems} from 'components/mdx/SidebarNavigation';
+import {getMDXItems} from 'components/SidebarNavigation';
 import rehypePlugins from 'components/mdx/rehypePlugins';
 import matter from 'gray-matter';
 import generateToc from 'mdast-util-toc';
@@ -72,16 +72,6 @@ function HTMLRenderer({data}: {data: HTMLData}) {
 }
 
 export default function MdxPage(props: Props) {
-  // const [isFeedbackOpen, setOpenFeedback] = useState<boolean>(false);
-
-  // const closeFeedback = () => {
-  //   setOpenFeedback(false);
-  // };
-
-  // const toggleFeedback = () => {
-  //   setOpenFeedback(!isFeedbackOpen);
-  // };
-
   const router = useRouter();
 
   // If the page is not yet generated, this shimmer/skeleton will be displayed
@@ -92,7 +82,6 @@ export default function MdxPage(props: Props) {
 
   return (
     <>
-      {/* <FeedbackModal isOpen={isFeedbackOpen} closeFeedback={closeFeedback} /> */}
       {props.type === PageType.MDX ? (
         <MDXRenderer data={props.data} />
       ) : (
@@ -188,7 +177,7 @@ export const getStaticProps: GetStaticProps = async ({params}) => {
     // 3. Extract table of contents from MDX
     const tree = remark().use(mdx).parse(content);
     const node = generateToc(tree, {maxDepth: 4});
-    const tableOfContents = getItems(node.map, {});
+    const tableOfContents = getMDXItems(node.map, {});
 
     // 4. Render MDX
     const mdxSource = await renderToString(content, {
