+++ docs/next/pages/_app.tsx
@@ -1,9 +1,11 @@
 import '/styles/fonts.css';
 import '/styles/globals.css';
 import '/styles/prism.css';
+import path from 'path';
 import {usePath} from 'util/usePath';
 
 import {PersistentTabProvider} from 'components/PersistentTabContext';
+import {collectHeadings, RightSidebar} from 'components/SidebarNavigation';
 import {DefaultSeo} from 'next-seo';
 import {AppProps} from 'next/app';
 import {useRouter} from 'next/router';
@@ -12,7 +14,7 @@ import * as React from 'react';
 import FeedbackModal from '../components/FeedbackModal';
 import Header from '../components/Header';
 import Sidebar from '../components/Sidebar';
-import {VersionedContentLayout, RightSidebar} from '../components/mdx/MDXRenderer';
+import {VersionedContentLayout} from '../components/mdx/MDXRenderer';
 import * as gtag from '../util/gtag';
 
 const BASE_URL = 'https://docs.dagster.io';
@@ -40,29 +42,6 @@ const DEFAULT_SEO = {
   },
 };
 
-// function collectHeadings(node, sections = []) {
-//   if (node) {
-//     if (node.name === 'Heading') {
-//       console.log(node);
-//       const title = node.children[0];
-
-//       if (typeof title === 'string') {
-//         sections.push({
-//           ...node.attributes,
-//           title,
-//         });
-//       }
-//     }
-
-//     if (node.children) {
-//       for (const child of node.children) {
-//         collectHeadings(child, sections);
-//       }
-//     }
-//   }
-
-//   return sections;
-// }
 interface Props {
   children: React.ReactNode;
   asPath: string;
@@ -136,7 +115,7 @@ const Layout: React.FC<Props> = ({asPath, children, pageProps}) => {
 
               <RightSidebar
                 editMode={editMode}
-                navigationItems={navigationItems}
+                navigationItemsForMDX={navigationItems}
                 githubLink={githubLink}
                 toggleFeedback={toggleFeedback}
               />
